requirements.txt
flask
mysql-connector-python
bs4
requests
lxml